fix: NaN DAC chain hole, coverage gaps, hardening, and tracker coast
Phase 1 — NaN hardening
- coordinate_mapper: isfinite + <=0 guards on both scale fields;
  DAC code range-checked as double before lround (lround(NaN) wrapped
  to 0, passing the integer check as full deflection)
- config_validator: new critical checks for dac_reference_voltage,
  finite bbox corners, finite/ordered galvo limits
- stereo_matcher: ctor sanitizes epipolar_tolerance_px
- 8 tests pinning each NaN / missing-boundary vector

Phase 2 — Coverage gaps + weak tests
- ConfigValidatorTest: boundary pairs for every critical bound
  (cooldown, settle, watchdog, grace, fps, dims, principal point,
  threshold, blob areas, max_blobs, epipolar); fixed principal-point
  inclusive off-by-one (cx == frame_width passed as 'inside')
- ControlLoopTest: target-loss glue (TRACKING->ARMED+clear,
  FIRING->COOLDOWN+abort), cooldown-exit re-arm, blanking window
  test now genuinely covers the blanking phase
- FiringControllerTest: disarm-abort asserts cooldown running
- Weak tests fixed: mapper DacValuesWithin12BitRange (off-axis sweep
  + variance proof), SPI backpressure (asserts ends on newest),
  E-stop ReleasedInitially (drives debounce), watchdog ordering
  (InSequence), jitter-test comment corrected
- New test_print.cpp (5 tests) + MCP4922 destructor re-centre test
- Real bug found+fixed in print.h: log_init stored stderr's flags
  after polluting them via the shared stdout/stderr open file
  description, so log_shutdown restored O_NONBLOCK back ON

Phase 3 — Hardening, tracker coast, docs
- detector: OpenCV exceptions -> fail closed; narrowing setter removed
- laser: firing_ only cleared on confirmed write; emergency_shutdown
  and destructor claim pin LOW only when achieved
- galvo: Y-write failure best-effort re-centres X
- camera: EINTR select loop uses absolute deadline
- print.h: resync flag never clobbered, tracked per-fd; log_shutdown
  made idempotent; log_init reads both original flags first
- main: RAII LogSessionGuard restores fd flags on every exit path
- tracker: coast through brief detection gaps using predict()
  (bounded by 100ms horizon, still gated by bounding box downstream);
  F_ member removed; per-pulse aim-point log restored
- AGENTS.md: SAFE_HALT claims, critical-bound list, logger wording
  in §5/§12, §7.1 table rows, §8 file list, §4.12 coast paragraph

Tests: 200 -> 229, Debug+Release clean, zero warnings.

+++ src/vision/detector.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "core/types.h"
-#include <optional>
 #include <vector>
 #include <cstdint>
 #include <cstddef>
@@ -23,10 +22,10 @@ public:
     [[nodiscard]] auto detect_blobs(const uint8_t* data, size_t size) const
         -> std::vector<Blob>;
 
-    void set_threshold(uint8_t threshold) { config_.threshold = threshold; }
-    [[nodiscard]] auto threshold() const -> uint8_t {
-        return static_cast<uint8_t>(config_.threshold);
-    }
+    // The config field is an int; narrowing through uint8_t here would wrap an
+    // out-of-range value silently at the call site.
+    void set_threshold(int threshold) { config_.threshold = threshold; }
+    [[nodiscard]] auto threshold() const -> int { return config_.threshold; }
     [[nodiscard]] auto width() const -> int { return width_; }
     [[nodiscard]] auto height() const -> int { return height_; }
